+++ m8rscript/Global.h
@@ -56,7 +56,7 @@ public:
     // Global has built-in properties. Handle those here
     virtual const Value property(ExecutionUnit*, const Atom&) const override;
     virtual Atom propertyName(ExecutionUnit*, uint32_t index) const override;
-    virtual size_t propertyCount(ExecutionUnit*) const override;
+    virtual uint32_t propertyCount(ExecutionUnit*) const override;
     
 protected:
     virtual bool serialize(Stream*, Error&, Program*) const override
